requirements.txt
hikvisionapi == 0.3.2
pywebview == 5.3.2
pyinstaller == 6.11.1